refactor(session): split pure title projection into title-projection.ts
title.ts mixed pure projection/validation (projectSessionTitle*,
sanitizeSessionTitle, titlePrompt) with provider transport, failure
classification, record-graph traversal, and the generation workflow. The
pure side moves to title-projection.ts verbatim; the store.ts barrel
re-exports the identical public surface. normalizeTitleUsage becomes an
export of the projection module (consumed by maybeGenerateSessionTitle)
without joining the barrel.

Refs #265
